+++ arm7/source/header.cpp
@@ -10,5 +10,9 @@ extern u8 __bss_size[];
 pload_header7_t gLoaderHeader
 {
     .entryPoint = (void*)&_start,
-    .apiVersion = PICO_LOADER_API_VERSION
+    .apiVersion = PICO_LOADER_API_VERSION,
+    .v4 =
+    {
+        .gameLanguage = PLOAD_GAME_LANGUAGE_AUTO
+    }
 };