+++ arm9/source/Arm7Patcher.h
@@ -9,10 +9,12 @@ public:
     /// @brief Applies arm7 patches using the given \p loaderPlatform.
     /// @param loaderPlatform The loader platform to use.
     /// @param cheatsLength The length of the cheats data, or zero when there are no cheats.
+    /// @param languageOverride The language to apply when the rom's arm7 loads the user settings,
+    ///                         or \c IPC_LANGUAGE_OVERRIDE_NONE for no override.
     /// @param cheatsPtr Pointer to where the cheats need to be stored, or \c nullptr when there are no cheats.
     /// @param bannerSavePathPtr Pointer to where the banner save path needs to be stored, or \c nullptr when there is no banner save.
     /// @param runInDSiMode \c true if the rom should be loaded in DSi mode, or \c false otherwise.
     /// @return A pointer to the patch space in IWRAM, or \c nullptr if the patches have been placed in main memory.
-    void* ApplyPatches(const LoaderPlatform* loaderPlatform, u32 cheatsLength,
+    void* ApplyPatches(const LoaderPlatform* loaderPlatform, u32 cheatsLength, u32 languageOverride,
         void*& cheatsPtr, char*& bannerSavePathPtr, bool runInDSiMode) const;
 };